JENKINS-60722 add maxDaysOld parameter to restrict age of multibranches
If there is no filter set and the configuration is changed or a new Jenkins is started (after an upgrade), all multi-branch pipeline jobs are executed at the same time. To reduce the amount of jobs running at the same time (especially with multiple projects, repositories and branches), a parameter is introduced, which restricts relevant branches, tags and pull requests according to the age of the head commit.